Rename debugger tests and add to execution
