--- backend.native/debugger-tests/build.gradle
@@ -1,19 +1,0 @@
-apply plugin: 'java'
-apply plugin: 'kotlin'
-
-dependencies {
-    compile "org.jetbrains.kotlin:kotlin-stdlib:$kotlinVersion"
-    compile project(path: ':backend.native', configuration: 'cli_bc')
-    compile "org.jetbrains.kotlin:kotlin-native-shared:$konanVersion"
-    compile 'junit:junit:4.12'
-}
-
-test {
-    testLogging { exceptionFormat = 'full' }
-
-    def dist = project.rootProject.file("dist").canonicalPath
-    systemProperties = [
-            'org.jetbrains.kotlin.native.home': dist
-    ]
-}
-test.dependsOn ':dist'
